Change all, of the values, to float and convert `GridSize` only in case with compatibility export

+++ osu.Game.Rulesets.Osu/Edit/OsuGridToolboxGroup.cs
@@ -214,14 +214,14 @@ namespace osu.Game.Rulesets.Osu.Edit
                 spacingSlider.ContractedLabelText = $"S: {spacing.NewValue:#,0.##}";
                 spacingSlider.ExpandedLabelText = $"Spacing: {spacing.NewValue:#,0.##}";
                 SpacingVector.Value = new Vector2(spacing.NewValue);
-                editorBeatmap.GridSize = (int)spacing.NewValue;
+                editorBeatmap.GridSize = spacing.NewValue;
             }, true);
 
             GridLinesRotation.BindValueChanged(rotation =>
             {
                 gridLinesRotationSlider.ContractedLabelText = $"R: {rotation.NewValue:#,0.##}";
                 gridLinesRotationSlider.ExpandedLabelText = $"Rotation: {rotation.NewValue:#,0.##}";
-                if (GridType.Value != PositionSnapGridType.Circle) editorBeatmap.GridRotation = (int)rotation.NewValue;
+                if (GridType.Value != PositionSnapGridType.Circle) editorBeatmap.GridRotation = rotation.NewValue;
             }, true);
 
             GridType.BindValueChanged(v =>
